Created a notebook, gui file and fixed np32int bug in the data field of the layers2d class.

+++ qim3d/process/layers2d.py
@@ -48,7 +48,7 @@ class Layers2d:
             if not isinstance(data, np.ndarray):
                 raise TypeError("Data must be a numpy.ndarray.")
         
-        self.data = data
+        self.data = data.astype(np.int32)
         self.is_inverted = is_inverted
         self.n_layers = n_layers
         self.delta = delta
@@ -72,7 +72,7 @@ class Layers2d:
         '''
         if not isinstance(data, np.ndarray):
             raise TypeError("Data must be a numpy.ndarray.")
-        self.data = data
+        self.data = data.astype(np.int32)
     
     def get_is_inverted(self):
         return self.is_inverted
